Silencing warnings one at a time

+++ lib/update/molecdyn/predictor/circular_buffer.h
@@ -292,7 +292,7 @@ namespace Chroma
 	unsigned int index = (start + i) % size_max;
 
 	x.resize(N5);
-	for(int s=0; s < N5; s++) { 
+	for(unsigned int s=0; s < N5; s++) { 
 	  x[s] = q[index][s];
 	}
       }
@@ -311,7 +311,7 @@ namespace Chroma
     {
       START_CODE();
 
-      if( e.size() != N5) { 
+      if( static_cast<int>(e.size()) != N5) { 
 	throw SizeMismatchException("Attempting to push std::vector of wrong size into circular buffer", N5, e.size());
       }
 
@@ -321,7 +321,7 @@ namespace Chroma
 	start = size_max -1; // end of list for first element
 	end = size_max -1;   // first element is also last element
 
-	for(int s = 0; s < N5; s++) {
+	for(unsigned int s = 0; s < N5; s++) {
 	  q[start][s]=e[s];          // store element
 	}
 
@@ -363,7 +363,7 @@ namespace Chroma
 	  } 
 	  
 	  // copy element to start
-	  for(int s = 0; s < N5; s++) { 
+	  for(unsigned int s = 0; s < N5; s++) { 
 	    q[start][s] = e[s];
 	  }
 	  // don't increase size
@@ -381,7 +381,7 @@ namespace Chroma
 	  }
 	  
 	  // Store element
-	  for(int s = 0; s < N5; s++) { 
+	  for(unsigned int s = 0; s < N5; s++) { 
 	    q[start][s] = e[s];
 	  }
 	  // Increase size count
